requirements.txt
torch
torchvision
timm
albumentations
opencv-python
numpy
pillow
pyyaml
tqdm
scikit-learn
matplotlib